% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accidentes2020_data.R
\docType{data}
\name{accidentes2020_data}
\alias{accidentes2020_data}
\title{Datos de accidentes de tráfico en 2020 en Madrid}
\format{
A data.table with 32427 rows and 17 variables:
\describe{
\item{num_expediente}{chr Número de expediente del accidente }
\item{fecha}{chr Fecha del accidente}
\item{hora}{chr Hora del accidente}
\item{localizacion}{chr Calle en la que se produce el accidente}
\item{numero}{chr Número de portal }
\item{distrito}{chr Distrito en el que se ubica la calle}
\item{tipo_accidente}{chr Tipología del accidente}
\item{estado_meteorológico}{chr Clima el día del accidente }
\item{tipo_vehiculo}{chr Tipo de vehiculo implicado en el accidente}
\item{tipo_persona}{chr Persona afectada por el accidente}
\item{rango_edad}{chr Rango de edad del accidentado}
\item{sexo}{chr Genero de la persona }
\item{lesividad}{chr }
\item{coordenada_x_utm}{num Coordenada X UTM }
\item{coordenada_y_utm}{num Coordenada Y UTM }
\item{positiva_alcohol}{chr ¿Positivo en la prueba de alcoholismo? }
\item{positiva_droga}{chr ¿Positivo en la prueba de drogas?}
}
}
\source{
?????
}
\usage{
accidentes2020_data
}
\description{
Datos de accidentes de tráfico en 2020 en Madrid
}
\keyword{datasets}
