% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activefires.R
\docType{data}
\name{activefires}
\alias{activefires}
\title{Datos de fuegos activos en EEUU y Latinoamérica entre 24 de Febrero y 3 de Marzo de 2022}
\format{
un ppp con 7639 puntos
}
\source{
https://firms.modaps.eosdis.nasa.gov/active_fire/
}
\description{
Datos de fuegos activos en EEUU y Latinoamérica entre 24 de Febrero y 3 de Marzo de 2022
}
\examples{
data(activefires)
plot(activefires)
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
\keyword{data}
