% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ayuntam.R
\docType{data}
\name{ayuntam}
\alias{ayuntam}
\title{Datos simulados sobre el signo político y la prestación de un determinado servicio.}
\format{
Un data.table con 8 filas y 2 variables cualitativas:
\describe{
\item{signo_gob}{chr Signo político del partido que gobierna el ayuntamiento.}
\item{serv}{chr Prestación o no de un deteminado servico.}
}
}
\source{
Datos cualitativos simulados.
}
\usage{
ayuntam
}
\description{
Datos simulados sobre el signo político que gobiera en 80 ayuntamientos y la prestación de un determinado servicio.
}
\keyword{datasets}
