% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ayuso.R
\docType{data}
\name{ayuso}
\alias{ayuso}
\title{Datos simulados sobre la opinión de la presidenta de la Comunidad de Madrid, Isabel D. Ayuso.}
\format{
Un data.table con 2795 filas y 2 variables cualitativas:
\describe{
\item{zona}{chr Zona de la Comunidad de Madrid donde se realizó la encuesta.}
\item{opinion}{chr Opinión del encuestado sobre Isabel D. Ayuso.}
}
}
\source{
Datos cualitativos simulados.
}
\usage{
ayuso
}
\description{
Datos simulados sobre la pecepción de la presidenta de la Comunidad de Madrid, Isabel D. Ayuso, en cuatro zonas de la región de Madrid.
}
\keyword{datasets}
