% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDR-package.R
\docType{package}
\name{CDR-package}
\alias{CDR}
\alias{CDR-package}
\title{CDR: Datos del libro Ciencia de Datos con R}
\description{
All the datasets used in the book "Ciencia de Datos con R".
}
\author{
\strong{Maintainer}: Itzcoatl Bueno \email{itzbueno@ucm.es} (\href{https://orcid.org/0000-0002-5629-1906}{ORCID})

}
\keyword{internal}
