% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\docType{data}
\name{cities}
\alias{cities}
\title{Asentamientos urbanos chilenos con mas de 2000 habitantes en 2002}
\format{
Objeto sf con 277 filas y 12 columnas
\describe{
\item{NAME}{chr Nombre resumido de la ciudad}
\item{REGION}{chr Región a la que pertenece según la codificación de 13 Regiones.}
\item{NAME_LONG}{chr Nombre completo de la ciudad}
\item{P30}{int Población en 1930}
\item{P40}{int Población en 1940}
\item{P52}{int Población en 1952}
\item{P60}{int Población en 1960}
\item{P70}{int Población en 1970}
\item{P82}{int Población en 1982}
\item{P92}{int Población en 1992}
\item{P02}{int Población en 2002}
}
}
\source{
Instituto nacional de Estadísticas de Chile (\url{http:://www.ine.cl})
}
\usage{
cities
}
\description{
Mapa de centuriones de las áreas urbanas chilenas con más de 2000 habitantes en 2002 usados en Vallone y Chasco (2020)
}
\references{
Vallone, A. Y Chasco, C. (2020). Spatiotemporal methods for analysis of urban system dynamics: an application to Chile. The Annals of Regional Science, 64(2):421–454.
}
\keyword{datasets}
