% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleveland.R
\docType{data}
\name{cleveland}
\alias{cleveland}
\title{Datos de pacientes que llegan a un hospital con dolor de pecho}
\format{
Un data.frame con 303 filas y 6 variables:
\describe{
\item{diag}{Factor diagnóstico de accidente coronario (0=no, 1=sí)}
\item{edad}{int edad en años cumplidos.}
\item{dep}{num depresión en el segmento ST inducida por ejercicio en relación al reposo.}
\item{sexo}{Factor sexo (0=mujer, 1=hombre)}
\item{tdolor}{Factor tipo de dolor (1=angina típica; 2=angina atípica; 3=dolor no anginal; 4=asintomático)}
\item{dhosp}{int número de días de hospitalización.}
}
}
\source{
Detrano, R. (1988). Heart Disease Databases. \url{https://archive.ics.uci.edu/ml/datasets/heart+Disease}.
}
\usage{
Cleveland.RData
}
\description{
Conjunto de datos de pacientes que llegan a un hospital con dolor de pecho, de los que se han registrado distintas características. Adaptado a partir de Detrano (1988)
}
\keyword{datasets}
