% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clothes.R
\docType{data}
\name{clothes}
\alias{clothes}
\title{Datos de reseñas y calificaciones de ropa de comercio electrónico para mujeres.}
\format{
Un data.frame con 23,486 filas y 10 variables:
\describe{
\item{ID}{chr Variable entera ordinal positiva entera que se refiere a la pieza específica que se está revisando.}
\item{Age}{int Variable entera ordinal positiva de la edad de los revisores.}
\item{Title}{chr Variable de cadena para el título de la reseña.}
\item{Review}{chr Variable de cadena para el cuerpo de revisión.}
\item{Rating}{int Variable entera ordinal positiva para el puntaje del producto otorgado por el cliente de 1 Peor a 5 Mejor.}
\item{Recommended}{int Variable entera ordinal positiva que indica dónde recomienda el cliente el producto donde 1 se recomienda, 0 no se recomienda.}
\item{Liked}{int Variable entera ordinal positiva que documenta el número de otros clientes que consideraron positivo este comentario.}
\item{Division}{chr Nombre categórico de la división de alto nivel del producto.}
\item{Dept}{chr Nombre categórico del nombre del departamento del producto.}
\item{Class}{chr Nombre categórico del nombre de clase del producto.}
}
}
\usage{
clothes
}
\description{
Womens Clothing E-Commerce Reviews. Datos de reseñas online de ropa de mujer de varios minoristas para predecir qué prendas tienen más probabilidades de recibir una nueva estrella Michelin
}
\keyword{datasets}
