% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumo_electrico_anual.R
\docType{data}
\name{consumoelectricoanual}
\alias{consumoelectricoanual}
\title{PIB per cápita y consumo eléctrico en España}
\format{
Un fichero csv, campos separados por punto y coma, con 30 filas y 3 variables:
\describe{
\item{Año}{num Año al que corresponden los datos.}
\item{PIB}{num PIB per cápita en España en miles de euros.}
\item{Consumo}{num Consumo eléctrico en España en miles de toneladas equivalentes de petróleo -ktep-.}
}
}
\source{
\itemize{
\item Consumo --> Web del IDAE (Instituto para la Diversificación y el Ahorro Energético): \url{https://sieeweb.idae.es/consumofinal/bal.asp?txt=Energ%EDa%20el%E9ctrica&tipbal=f&rep=1}
\item PIB --> Web del diario Expansión (\url{https://datosmacro.expansion.com/pib/espana})
}
}
\usage{
ConsumoElectricoAnual.csv
}
\description{
Datos anuales de PIB per cápita y consumo eléctrico en España entre 1990 y 2019
}
\keyword{datasets}
