% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_data_valencia.R
\docType{data}
\name{crime_data_valencia}
\alias{crime_data_valencia}
\title{Data of crimes in the city of Valencia (Spain)}
\format{
A data.table with 5338 rows and 6 variables:
\describe{
\item{crime_id}{chr Identificador del crimen}
\item{crime_type}{chr Tipo de crimen cometido}
\item{crime_date}{chr Fecha en la que se cometió el crimen}
\item{year}{int Año en el que se produce el crimen}
\item{crime_lat}{num Latitud}
\item{crime_lon}{num Longitud}
}
}
\source{
?????
}
\usage{
crime_data_valencia
}
\description{
Crimes are classified by type ubicated in Valencia during 2010
}
\keyword{datasets}
