% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_elecciones.R
\docType{data}
\name{datos_elecciones}
\alias{datos_elecciones}
\title{Resultados electorales en Andalucia}
\format{
Un tibble de 113.161 filas y 13 variables:
\describe{
\item{convocatoria}{Date Fecha de la convocatoria electoral}
\item{codigo_secc}{chr Código de la sección censal}
\item{censo}{num ...}
\item{part}{num Porcentaje de participación}
\item{votantes}{num Número de votantes en la seccion censal}
\item{blancos}{num Número de votos en blanco}
\item{nulos}{num Número de votos nulos}
\item{validos}{num Número de votos validos}
\item{partido}{num Partido político más votado en la sección censal}
\item{votos_partido}{num Votos recibidos por el partido ganador de la sección censal}
\item{votos_partido_pc}{num ...}
\item{codigo_num}{chr ...}
\item{bloque}{chr Orientación politica del bloque ganador}
}
}
\source{
Instituto Nacional de Estadística (INE)
}
\usage{
datos_elecciones
}
\description{
Resultados electorales en las elecciones de Andalucía
}
\keyword{datasets}
