% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_observacional.R
\docType{data}
\name{datos_observacional}
\alias{datos_observacional}
\title{Datos observacional}
\format{
Un data.frame con 100 filas y 8 variables cualitativas:
\describe{
\item{ID}{num identificador del paciente}
\item{fecha_hospitalizacion}{date fecha de hospitalización del paciente}
\item{sexo}{chr sexo al nacer}
\item{edad}{num edad al diagnóstico}
\item{comorbilidades}{chr presencia de una o más comorbilidades}
\item{fecha_alta}{chr fecha de alta del paciente}
\item{exitus}{num exitus}
\item{fecha_exitus}{date fecha de exitus del paciente}
}
}
\usage{
datos_observacional
}
\description{
Datos observacional
}
\keyword{datasets}
