% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_supervivencia.R
\docType{data}
\name{datos_supervivencia}
\alias{datos_supervivencia}
\title{Datos supervivencia}
\format{
Un data.frame con 301 filas y 7 variables:
\describe{
\item{id}{num identificador del paciente}
\item{EXITUS_TIME}{num tiempo desde diagnóstico COVID-19 hasta exitus}
\item{DIAG_COVID}{num diagnóstico de COVID-19}
\item{EXITUS}{num exitus}
\item{N_COMORBIDITIES}{num número de comorbilidades}
\item{SEX}{chr sexo al nacer}
\item{EDAD}{num edad al diagnóstico}
}
}
\usage{
datos_supervivencia
}
\description{
Datos supervivencia
}
\keyword{datasets}
