% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaracion.R
\docType{data}
\name{declaracion}
\alias{declaracion}
\title{Declaración institucional del presidente del gobierno.}
\format{
Datos tipo character.
}
\source{
\url{https://www.lamoncloa.gob.es/presidente/actividades/Paginas/2020/130320-sanchez-declaracio.aspx}
}
\usage{
declaracion
}
\description{
Declaración institucional del presidente del gobierno sobre el estado de alarma en la crisis del coronavirus, dada en La Moncloa el 13 de marzo de 2020.
}
\keyword{datasets}
