% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_fecha_venta.R
\docType{data}
\name{dim_fecha_venta}
\alias{dim_fecha_venta}
\title{Tabla con fechas de ventas}
\format{
Un data.frame con 94 filas y 2 variables.
\describe{
\item{IDE_FECHA_VENTA}{chr Código identificador de las fechas en las que se han producido ventas con formato AAAA-MM-DD (año-mes-día)}
\item{DES_FECHA_VENTA}{chr Descripción de la fecha de la venta}
}
}
\usage{
dim_fecha_venta
}
\description{
Tabla que contiene las distintas fechas en las que se han producido ventas en la empresa.
}
\keyword{datasets}
