% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_nivel_educacion_cliente.R
\docType{data}
\name{dim_nivel_educacion_cliente}
\alias{dim_nivel_educacion_cliente}
\title{Tabla con los niveles de educación de los clientes}
\format{
Un data.frame con 94 filas y 2 variables.
\describe{
\item{IDE_NIVEL_EDUCACION_CLIENTE}{int Código identificador del nivel de educación del cliente}
\item{DES_NIVEL_EDUCACION_CLIENTE}{chr Descripción del nivel de educación del cliente}
}
}
\usage{
dim_nivel_educacion_cliente
}
\description{
Los posibles niveles educacaciones de los clientes de nuestra empresan se encuentra en esta tabla.
}
\keyword{datasets}
