% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev.R
\docType{data}
\name{elev}
\alias{elev}
\title{Datos de la provincia de Toledo}
\format{
A RasterLayer with 32 row, 75 variables and 2400 cells
\describe{
}
}
\source{
?????
}
\usage{
elev
}
\description{
Datos de la provincia de Toledo
}
\keyword{datasets}
