% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elon_musk.R
\docType{data}
\name{elon_musk}
\alias{elon_musk}
\title{Tweets de Elon Musk}
\format{
Un data.frame con 3197 filas y 7 variables:
\describe{
\item{id_str}{num Identificador del tweet}
\item{created_at}{POSIXct Fecha de creación del tweet}
\item{full_text}{chr Texto del tweet}
\item{tipo_tweet}{chr Tipo de tweet
\itemize{
\item original: tweet original
\item quoted: tweet que cita a otro tweet
\item reply: tweet que responde a otro tweet
\item retweeted: tweet que retuitea a otro tweet}
\item{retweet_count}{num Número de retweets recibidos por el tweet}
\item{favorite_count}{num Número de favoritos recibidos por el tweet}
}
}
}
\source{
Twitter
}
\usage{
elon_musk
}
\description{
Tweets obtenidos con la API Standard de Twitter V1.1. el día 22 de diciembre de 2022.
}
\keyword{datasets}
