% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_declaracion.R
\docType{data}
\name{EN_declaracion}
\alias{EN_declaracion}
\title{Declaración institucional del presidente del gobierno en inglés.}
\format{
Datos tipo character.
}
\usage{
EN_declaracion
}
\description{
Traducción automática al inglés de la Declaración institucional del presidente del gobierno sobre el estado de alarma en la crisis del coronavirus, dada en La Moncloa el 13 de marzo de 2020.
}
\keyword{datasets}
