% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_venta.R
\docType{data}
\name{fac_venta}
\alias{fac_venta}
\title{Tabla de compras realizadas por clientes}
\format{
Un data.frame con 11354 filas y 8 variables.
\describe{
\item{IDE_VENTA}{int Código identificador único de la venta al cliente}
\item{IDE_PEDIDO}{int Código identificador del pedido del cliente}
\item{IDE_CLIENTE}{int Código identificador del cliente}
\item{IDE_FECHA_VENTA}{chr Código identificador de la fecha de la venta}
\item{IDE_PRODUCTO}{chr Código identificador del producto de la venta}
\item{IDE_VENDEDOR}{int Código identificador del vendedor de esa venta}
\item{NUM_UNIDADES}{int Unidades del producto de esa venta}
\item{NUM_IMP}{int Importe neto de todas las unidades del producto de esa venta}
}
}
\usage{
fac_venta
}
\description{
Las distintas compras que han hecho los clientes están en la tabla con los hechos correspondientes.
}
\keyword{datasets}
