% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/festivos.R
\docType{data}
\name{festivos}
\alias{festivos}
\title{Datos de festivos en España desde 1995 hasta 2044}
\format{
Un csv con 500 filas y 4 variables:
\describe{
\item{ds}{date Fecha del festivo.}
\item{holiday}{chr Nombre del festivo.}
\item{country}{chr País del festivo.}
\item{year}{num Año del festivo}
}
}
\usage{
festivos
}
\description{
Fechas de los festivos en España con el nombre exacto de cada uno de ellos
}
\keyword{datasets}
