% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdpmap.R
\docType{data}
\name{gdpmap}
\alias{gdpmap}
\title{Municipios (NUTS 5) que forman parte de las áreas urbanas de España}
\format{
Objeto sp com 122 filas y 13 columnas
\describe{
\item{POLYID}{num Identificador}
\item{CODITO}{num Código de municipio INE}
\item{MUNICIPIO}{chr Nombre del municipio}
\item{PROVINCIA}{chr Nombre de la provincia}
\item{CCAA}{chr Nombre de la Comunidad Autónoma}
\item{CP}{num Variable dicotómica que mide el crecimiento del capital per cápita por provincias en el periodo 1985-1998; toma el valor 1 para las ciudades situadas en aquellas regiones españolas que han registrado un crecimiento de la renta superior a la media y 0 para el resto. crecimiento de la renta y 0 para el resto.}
\item{XCOOR}{num Coordenada X del centroide}
\item{YCOOR}{num Coordenada Y del centroide}
\item{LPGH}{num Logaritmo del PIB per cápita de España a nivel municipal en 2001}
\item{LPGH85}{num Logaritmo del PIB per cápita de España a nivel municipal en 2001}
\item{BANK}{num Tasa de crecimiento del número de bancos y cajas de ahorro per cápita registrado en una ciudad durante el periodo 1991-2001}
\item{UNI01}{num La evolución de la tasa de desempleo de las ciudades en el periodo 1991-2001}
\item{PAT00}{num Número de patentes per cápita por ciudades referidos al año 2000}
}
}
\source{
Varias, para detalles refiérase a Mella y Chasco (2006)
}
\usage{
gdpmap
}
\description{
Datos utilizados en Mella y Chasco (2006)
}
\references{
José M. Mella-Márquez and Coro Chasco-Yrigoyen, 2006. "Urban Growth and Territorial Dynamics: A Spatial-Econometric Analysis of Spain," Chapters, in: Aura Reggiani & Peter Nijkamp (ed.), Spatial Dynamics, Networks and Modelling, chapter 13, Edward Elgar Publishing.
}
\keyword{datasets}
