% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_toledo.R
\docType{data}
\name{hosp_toledo}
\alias{hosp_toledo}
\title{Datos de hospitales de Toledo (España)}
\format{
A sf with 30 rows and 26 variables:
\describe{
\item{country}{}
\item{city}{}
\item{cap_beds}{}
\item{emergency}{}
\item{ref_date}{}
\item{lon}{}
\item{house_number}{}
\item{pub_date}{}
\item{cap_rooms}{}
\item{street}{}
\item{tel}{}
\item{id}{}
\item{facility_type}{}
\item{list_specs}{}
\item{lat}{}
\item{email}{}
\item{hospital_name}{}
\item{cc}{}
\item{public_private}{}
\item{comments}{}
\item{postcode}{}
\item{url}{}
\item{cap_prac}{}
\item{site_name}{}
\item{geo_qual}{}
\item{geometry}{}
}
}
\source{
?????
}
\usage{
hosp_toledo
}
\description{
Datos de hospitales de Toledo (España)
}
\keyword{datasets}
