% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marvel_edges.R
\docType{data}
\name{marvel_edges}
\alias{marvel_edges}
\title{Relación entre los personajes pertenecientes al Universo Marvel.}
\format{
Un data.table con 98 filas y 2 variables
\describe{
\item{hero_from}{chr Nombre del personaje perteneciente al Universo Marvel origen de la relación.}
\item{hero_to}{chr Nombre del personaje perteneciente al Universo Marvel destino de la relación.}
}
}
\source{
Elaboración propia.
}
\usage{
marvel_edges
}
\description{
Contiene dos columnas formateadas para representar la red social del Universo Marvel, donde la primera columna es el nombre de un personaje del Universo cinematografico Marvel y la segunda el nombre de otro con el que coincide en alguna pelicula, representado cada relación una arista entre dos nodos.
}
\keyword{datasets}
