% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marvel_nodes.R
\docType{data}
\name{marvel_nodes}
\alias{marvel_nodes}
\title{Nombres de distintos personajes pertenecientes al Universo Marvel.}
\format{
Un data.table con 6421 filas y 1 variable
\describe{
\item{hero}{chr Nombre del personaje perteneciente al Universo Marvel.}
}
}
\source{
\url{http://syntagmatic.github.io/exposedata/marvel/}
}
\usage{
marvel_nodes
}
\description{
Contiene una columna con los nombres de los super héroes de Marvel. Cada héroe es un nodo de la red social.
}
\keyword{datasets}
