% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_conf_model.R
\docType{data}
\name{md_conf_model}
\alias{md_conf_model}
\title{Configuración de modelos de ML}
\format{
Una lista de 6 elementos.
}
\usage{
md_conf_model
}
\description{
Lista con la configuración y datos básicos para construir y validar todos los modelos a desarrollar
}
\keyword{datasets}
