% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medellin.R
\docType{data}
\name{medellin}
\alias{medellin}
\title{Datos de crímenes en medellín durante 2005}
\format{
un ppp con 910 puntos
}
\description{
Datos de crímenes en medellín durante 2005
}
\examples{
data(medellin)
plot(medellin)
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
\keyword{data}
