% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipios.R
\docType{data}
\name{municipios}
\alias{municipios}
\title{Datos de municipios de España}
\format{
A sf with 8212 rows and 8 variables:
\describe{
\item{codauto}{factor Código de CCAA}
\item{ine.ccaa.name}{factor Nombre de la CCAA}
\item{cpro}{factor Código de provincia}
\item{ine.prov.name}{factor Nombre de la provincia}
\item{cmun}{factor Código de municipio}
\item{name}{factor Nombre del municipio}
\item{LAU_CODE}{factor}
\item{geom}{sfc_MULTIPOLYGON}
}
}
\source{
?????
}
\usage{
municipios
}
\description{
Datos de municipios de España
}
\keyword{datasets}
