% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepal.R
\docType{data}
\name{nepal}
\alias{nepal}
\title{Datos de incendio forestal en nepal durante 2016}
\format{
un ppp con 5757 puntos
}
\description{
Datos de incendio forestal en nepal durante 2016
}
\examples{
data(nepal)
plot(nepal)
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
\keyword{data}
