% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempmax_data.R
\docType{data}
\name{tempmax_data}
\alias{tempmax_data}
\title{Datos de temperatura máxima del aire registradas en España en 2022}
\format{
Una lista que contiene 6 objetos con distintos formatos: ESP, ESP_utm, grd_sf, grd_sp, temp_max_utm_sf, temp_max_utm_sp
\describe{
\item{ESP}{Objeto "sfc_MULTIPOLYGON" "sfc" descargado de la librería mapSpain. Geodetic CRS:  WGS 84.}
\item{ESP_utm}{Objeto \code{ESP} con proyección EPSG:25830, UTM ETRS89 Huso 30 N haciendo \code{ESP_utm <- st_transform(ESP, 25830)}.}
\item{grd_sf}{Objeto "sfc_POINT" "sfc" creado para utilizarse como malla de España.}
\item{grd_sp}{Objeto \code{grd_sf} convertido a un objeto \code{sp} (\code{SpatialPixels}) necesario para llevar a cabo la interpolación, que debe hacerse con un objeto sp/raster.}
\item{temp_max_utm_sf}{Objeto "sf", "data.frame", con los datos de temperatura máxima (grados centígrados) del aire registradas el España y sus coorrespondientes coordenadas (\code{longitud}y \code{latitud}) con proyeccion con proyección EPSG:25830.}
\item{temp_max_utm_sp}{Objeto \code{temp_max_utm_sf} convertido a \code{SpatialPointsDataFrame}.}
}
}
\source{
Agencia Estatal de Meteorología (AEMET) y libería mapSpain.
}
\usage{
tempmax_data
}
\description{
Datos de temperatura máxima del aire registrados en España por las estaciones meteorológicas de la Agencia Estatal de Meteorología (AEMET) el día 6 de agosto de 2022, día record en dicho año. Estos datos han sido descargados usando la librería climaemet. Además, se proporcionan los objetos espaciales relativos al contorno de España una malla para interpolación.
}
\keyword{datasets}
