% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempmin_data.R
\docType{data}
\name{tempmin_data}
\alias{tempmin_data}
\title{Datos de temperaturas}
\format{
A data.table with 1066 rows and 5 variables:
\describe{
\item{fecha}{chr Fecha}
\item{indicativo}{chr }
\item{tmin}{num Temperatura mínima}
\item{longitud}{num longitud}
\item{latitud}{num latitud}
}
}
\source{
?????
}
\usage{
tempmin_data
}
\description{
Datos de temperaturas
}
\keyword{datasets}
