% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tol_prov.R
\docType{data}
\name{Tol_prov}
\alias{Tol_prov}
\title{Archivo de datos raster de la provincia de Toledo}
\format{
Un sf con 1 fila y 24 variables:
\describe{
\item{codauto}{factor Código de la CCAA}
\item{cpro}{factor Código de la provincia}
\item{iso2.prov.code}{factor}
\item{nuts.prov.code}{factor}
\item{ine.prov.name}{factor}
\item{iso2.prov.name.es}{factor}
\item{iso2.prov.name.ca}{factor}
\item{iso2.prov.name.ga}{factor}
\item{iso2.prov.name.eu}{factor}
\item{cldr.prov.name.en}{factor}
\item{cldr.prov.name.es}{factor}
\item{cldr.prov.name.ca}{factor}
\item{cldr.prov.name.ga}{factor}
\item{cldr.prov.name.eu}{factor}
\item{prov.shortname.en}{factor}
\item{prov.shortname.es}{factor}
\item{prov.shortname.ca}{factor}
\item{prov.shortname.ga}{factor}
\item{prov.shortname.eu}{factor}
\item{nuts2.code}{factor}
\item{nuts2.name}{factor}
\item{nuts1.code}{factor}
\item{nut1.name}{factor}
\item{geom}{sfc_POLYGON}
}
}
\source{
Instituto Geográfico Nacional (IGN). \url{https://www.ign.es/web/ign/portal}
}
\usage{
Tol_prov
}
\description{
Datos espaciales (polígonos) de la provincia de Toledo. Se ha extraído del Instituto Geográfico Nacional (IGN) usando el paquete mapSpain. Contiene varias columnas con el código correspondiente a la provincia según diferentes esquemas de clasificación (INE, NUTS, ISO) así como diferentes denominaciones en varios idiomas (inglés, castellano, catalán, gallego, euskera).
}
\keyword{datasets}
