% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toledo.R
\docType{data}
\name{toledo}
\alias{toledo}
\title{Geometría de la provincia de Toledo}
\format{
Un sf con 1 fila y 8 variables:
\describe{
\item{codauto}{factor Código de la CCAA.}
\item{ine.ccaa.name}{factor Nombre de la CCAA.}
\item{cpro}{factor Código de la provincia.}
\item{ine.prov.name}{factor Nombre de la provincia.}
\item{cmun}{factor Código del municipio.}
\item{name}{factor Nombre del municipio (Toledo).}
\item{LAU_CODE}{factor}
\item{geom}{sfc_POLYGON}
}
}
\source{
Instituto Geográfico Nacional (IGN). \url{https://www.ign.es/web/ign/portal}
}
\usage{
toledo
}
\description{
Datos de la ciudad de Toledo.
}
\keyword{datasets}
