% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valencia.R
\docType{data}
\name{valencia}
\alias{valencia}
\title{Datos de accidentes de tráfico en la ciudad de valencia durante 2020}
\format{
un lpp con 4222 puntos
}
\description{
Datos de accidentes de tráfico en la ciudad de valencia durante 2020
}
\examples{
data(valencia)
plot(valencia)
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
\keyword{data}
