% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_facebook.R
\docType{data}
\name{datos_facebook}
\alias{datos_facebook}
\title{Datos simulados sobre las relaciones de una cuenta en la red social Facebook.}
\format{
Un data.frame con 1999 filas y 2 variable:
\describe{
\item{P1}{int Representa con un ID a la persona origen de la relación.}
\item{P2}{int Representa con un ID a la persona destino de la relación.}
}
}
\source{
Estos datos, en su forma real, se puden obtener de \url{https://developers.facebook.com} mediante una cuenta de desarrollador, para ello es necesario activar desde esta web la API. Se puede obtener más información desde \url{https://developers.facebook.com/docs}.
}
\usage{
datos_facebook
}
\description{
Este fichero representa las relaciones entre unas personas, reprentadas por un id ficticio, que pertenecen a la red social Facebook. Se pueden observar dos columnas donde cada una de ellas es el id de una persona que tiene relacion con el id de otra. Estos datos, en su forma real para cada cuenta, se pueden obtener de \url{https://developers.facebook.com} mediante una cuenta de desarrollador, para ello es necesario activar desde esta web la API. Se puede obtener más información desde https://developers.facebook.com/docs.
}
\keyword{datasets}
