% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_retail.R
\docType{data}
\name{datos_retail}
\alias{datos_retail}
\title{Datos para la segmentación de clientes de una empresa de venta minorista}
\format{
Un data.table con 200 000 filas y 4 variables
\describe{
\item{id_ticket}{chr  Identificador de ticket para una de las ventas}
\item{fecha}{date Abarca las fechas de las transaciones en el en el periodo de estudi octubre 2019 a julio de 2022}
\item{importe_venta}{dbl Recoge el importe neto de cada transacción de venta en euros, sin incluir impuestos ni otros costes adicionales que puedan distorsionar la comparativa entre los clientes objeto de la segmentación}
\item{codigo_socio}{chr Contiene el código identificador único de cada cliente vinculado a cada transacción de compra}
}
}
\source{
Datos anonimizados
}
\usage{
datos_retail.csv
}
\description{
Datos de las transaciones de compra para el periodo de estudio octubre 2019 a julio de 2022
}
\keyword{datasets}
