% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev.R
\docType{data}
\name{elev}
\alias{elev}
\title{Datos de elevación de la provincia de Toledo en metros}
\format{
Un RasterLayer con 32 filas, 75 variables y 2400 celdas.
\describe{
\item{x}{}
\item{y}{}
}
}
\source{
DIVA-GIS \url{http://www.diva-gis.org/gdata} mediante el uso del paquete R raster.
}
\usage{
elev
}
\description{
Datos de la provincia de Toledo
}
\keyword{datasets}
