% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renta_municipio_data.R
\docType{data}
\name{renta_municipio_data}
\alias{renta_municipio_data}
\title{Datos de renta per capita por municipio (en euros)}
\format{
A data.table with 55273 rows and 6 variables:
\describe{
\item{Unidad}{chr Identificador del municipio}
\item{\code{2019}}{int Renta per cápita en 2019}
\item{\code{2018}}{int Renta per cápita en 2018}
\item{\code{2017}}{int Renta per cápita en 2017}
\item{\code{2016}}{int Renta per cápita en 2016}
\item{\code{2015}}{int Renta per cápita en 2015}
}
}
\source{
?????
}
\usage{
renta_municipio_data
}
\description{
Datos de renta per capita por municipio (en euros)
}
\keyword{datasets}
