% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tajo.R
\docType{data}
\name{tajo}
\alias{tajo}
\title{Data of Tajo river}
\format{
A sf with 16 rows and 11 variables:
\describe{
\item{id_rio}{factor }
\item{pers_hidro}{int }
\item{orig_hidro}{int }
\item{ubicacion}{int }
\item{t_rio}{int }
\item{rotulo}{factor }
\item{fecha_alta}{date }
\item{fecha_baja}{date }
\item{st_length_}{num }
\item{NOM_RIO}{factor }
\item{geometry}{sfc_LINESTRING}
}
}
\source{
????
}
\usage{
tajo
}
\description{
Loads a sf object representing Tajo river (Spain)
}
\keyword{datasets}
