% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tol_prov.R
\docType{data}
\name{Tol_prov}
\alias{Tol_prov}
\title{Datos de la provincia de Toledo}
\format{
A sf with 1 row and 24 variables:
\describe{
\item{codauto}{factor Código de la CCAA}
\item{cpro}{factor Código de la provincia}
\item{iso2.prov.code}{factor}
\item{nuts.prov.code}{factor}
\item{ine.prov.name}{factor}
\item{iso2.prov.name.es}{factor}
\item{iso2.prov.name.ca}{factor}
\item{iso2.prov.name.ga}{factor}
\item{iso2.prov.name.eu}{factor}
\item{cldr.prov.name.en}{factor}
\item{cldr.prov.name.es}{factor}
\item{cldr.prov.name.ca}{factor}
\item{cldr.prov.name.ga}{factor}
\item{cldr.prov.name.eu}{factor}
\item{prov.shortname.en}{factor}
\item{prov.shortname.es}{factor}
\item{prov.shortname.ca}{factor}
\item{prov.shortname.ga}{factor}
\item{prov.shortname.eu}{factor}
\item{nuts2.code}{factor}
\item{nuts2.name}{factor}
\item{nuts1.code}{factor}
\item{nut1.name}{factor}
\item{geom}{sfc_POLYGON}
}
}
\source{
?????
}
\usage{
Tol_prov
}
\description{
Datos de la provincia de Toledo
}
\keyword{datasets}
