% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toledo.R
\docType{data}
\name{toledo}
\alias{toledo}
\title{Datos la provincia de Toledo}
\format{
A sf with 1 row and 8 variables:
\describe{
\item{codauto}{factor Código de la CCAA}
\item{ine.ccaa.name}{factor Nombre de la CCAA}
\item{cpro}{factor Código de la provincia}
\item{ine.prov.name}{factor Nombre de la provincia}
\item{cmun}{factor Código del municipio}
\item{name}{factor Nombre del municipio}
\item{LAU_CODE}{factor}
\item{geom}{sfc_POLYGON}
}
}
\source{
?????
}
\usage{
toledo
}
\description{
Datos la provincia de Toledo
}
\keyword{datasets}
